requirements.txt
Django==3.2.15
djangorestframework==3.14.0
django-environ==0.9.0
django-filter==21.1
django-autoslug==1.9.8
Pillow==9.2.0
django-phonenumber-field==7.0.0
phonenumbers==8.12.56
psycopg2-binary==2.9.3
flake8==5.0.4
black==22.8.0
isort==5.10.1
djoser==2.1.0
djangorestframework-simplejwt==4.8.0
PyJWT==2.5.0
pytest-django==4.5.2
pytest-factoryboy==2.5.0
Faker==14.2.1
pytest-cov==3.0.0
drf-yasg